backend/requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
dj-database-url==2.1.0
PyJWT==2.11.0
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
